requirements.txt
torch
torchvision
ftfy
regex
tqdm
pillow
git+https://github.com/openai/CLIP.git
requests
pandas
scikit-learn